Move crossplane into subpackage
