+++ hack/generator/pkg/codegen/pipeline/crossplane_add_embedded_resource_spec.go
@@ -3,7 +3,7 @@
  * Licensed under the MIT license.
  */
 
-package codegen
+package pipeline
 
 import (
 	"context"
@@ -11,15 +11,14 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/Azure/azure-service-operator/hack/generator/pkg/astmodel"
-	"github.com/Azure/azure-service-operator/hack/generator/pkg/codegen/pipeline"
 )
 
 var CrossplaneRuntimeV1Alpha1Package = astmodel.MakeExternalPackageReference("github.com/crossplane/crossplane-runtime/apis/core/v1alpha1")
 
-// addCrossplaneEmbeddedResourceSpec puts an embedded runtimev1alpha1.ResourceSpec on every spec type
-func addCrossplaneEmbeddedResourceSpec(idFactory astmodel.IdentifierFactory) pipeline.Stage {
+// AddCrossplaneEmbeddedResourceSpec puts an embedded runtimev1alpha1.ResourceSpec on every spec type
+func AddCrossplaneEmbeddedResourceSpec(idFactory astmodel.IdentifierFactory) Stage {
 
-	return pipeline.MakeStage(
+	return MakeStage(
 		"addCrossplaneEmbeddedResourceSpec",
 		"Adds an embedded runtimev1alpha1.ResourceSpec to every spec type",
 		func(ctx context.Context, types astmodel.Types) (astmodel.Types, error) {
